+++ CMSIS/DSP/Source/StatisticsFunctions/arm_min_q15.c
@@ -66,7 +66,7 @@ void arm_min_q15(
     mve_pred16_t p0;
 
 
-    indexVec = vidupq_u16(0, 1);
+    indexVec = vidupq_u16((uint32_t)0, 1);
     curExtremIdxVec = vdupq_n_u16(0);
 
     blkCnt = blockSize >> 3;
